examples/SlideReader/requirements.txt
python_pptx==0.6.17
pyttsx3==2.7
